requirement.txt
streamlit
firebase-admin
Pillow
pandas
opencv-python-headless
numpy
requests